Consolidate mount test into betterproto experimental module tests setup
